Move type BlockItem to types.ts
